[VAS-1070] moved class
